backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
chromadb==0.5.20
google-generativeai==0.3.1
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pandas==2.1.3
numpy==1.25.2
aiofiles==23.2.1
